+++ kipper/cli/src/index.ts
@@ -14,7 +14,7 @@ export * as help from "./help";
 // eslint-disable-next-line no-unused-vars
 export const name = "@kipper/cli";
 // eslint-disable-next-line no-unused-vars
-export const version = "0.8.2";
+export const version = "0.8.3";
 // eslint-disable-next-line no-unused-vars
 export const author = "Luna Klatzer";
 // eslint-disable-next-line no-unused-vars